Property::key same type as Property::hash Collision cleanups some const
git-svn-id: http://svn.sagaengine.com/sagaengine/trunk@622 65b84acb-c606-4e19-8676-f724b2686c9c

+++ core/src/sim/react/CollisionManager.hpp
@@ -74,7 +74,7 @@ namespace se_core {
 		CollisionGrid** gridPool_;
 
 		static const int MAX_CONTACTS = 1024;
-		int contactCount_;
+		short contactCount_;
 		Contact* contactList_;
 
 	};
